Simplify example
